moved menu into seperate activity
