+++ app/src/main/AndroidManifest.xml
@@ -17,7 +17,8 @@
             </intent-filter>
         </activity>
         <activity android:name=".HomeActivity" />
-        <activity android:name=".DetailActivity"></activity>
+        <activity android:name=".DetailActivity" />
+        <activity android:name=".MenuActivity"></activity>
     </application>
 
 </manifest>
