--- app/src/main/java/Model/BetterActionBar.java
@@ -1,8 +1,0 @@
-package Model;
-
-/**
- * Created by brandonholderman on 10/26/17.
- */
-
-public class BetterActionBar {
-}
