+++ app/src/main/java/com/example/louis/myapplication/MenuActivity.java
@@ -1,0 +1,39 @@
+package com.example.louis.myapplication;
+
+import android.support.v4.widget.DrawerLayout;
+import android.support.v7.app.ActionBarDrawerToggle;
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+import android.view.MenuItem;
+
+import butterknife.BindView;
+import butterknife.ButterKnife;
+
+public class MenuActivity extends AppCompatActivity {
+
+    private ActionBarDrawerToggle mToggle;
+
+    @BindView(R.id.drawer)
+    DrawerLayout mDrawerLayout;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.navigation_drawer);
+        ButterKnife.bind(this);
+
+        mToggle = new ActionBarDrawerToggle(this, mDrawerLayout, R.string.open, R.string.close);
+        mDrawerLayout.addDrawerListener(mToggle);
+        mToggle.syncState();
+        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        if (mToggle.onOptionsItemSelected(item)) {
+            return true;
+        }
+
+        return super.onOptionsItemSelected(item);
+    }
+}
